Don't allow FSX scripts to reference running framework
This causes issues in cases where the framework running FSLS
is incomplete or older than the other assemblies in the system
framework. This usually manifests as errors saying that various
assemblies cannot be found.

Instead, try to replace references to FSLS assemblies with the
system ones and ignore ones which don't exist in the system
framework.

(Rest of the differences are due to whitespace cleanup)